fixed issues on tester

+++ api/views.py
@@ -8,7 +8,7 @@ from api.models import Person
 from api.serializers import PersonSerializer
 
 # Create your views here.
-class BasePersonView(APIView):
+class BaseView(APIView):
     serializer = PersonSerializer
     queryset = Person.objects.all
 
@@ -36,12 +36,11 @@ class BasePersonView(APIView):
         # Return errors if the data is invalid
         return Response(new_person.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class ReadUpdateDeleteView(APIView):
+class CRUDView(APIView):
     serializer = PersonSerializer
 
-    def get(self, request, user_id=None, *args, **kwargs):
+    def get(self, request, user_id=None, name=None, *args, **kwargs):
         # Read info aobut a person using id or name
-        name = request.query_params.get('name')
         if name:
             person = get_object_or_404(Person, name=name)
         else:
@@ -50,9 +49,9 @@ class ReadUpdateDeleteView(APIView):
         
         return Response(serialized_person.data, status=status.HTTP_200_OK)
         
-    def put(self, request, user_id, *args, **kwargs):
+    def put(self, request, user_id=None, name=None, *args, **kwargs):
         # Update info aobut a person using id or name
-        name = request.query_params.get('name')
+        print(request.data.get('name'))
         if name:
             person = get_object_or_404(Person, name=name)
         else:
@@ -61,7 +60,7 @@ class ReadUpdateDeleteView(APIView):
 
         # Update and return an existing Person object only if there's no person with the new name.
         if update_data.is_valid():
-            if not Person.objects.filter(name=request.data['name'] or request.query_params['name']).exists():
+            if not Person.objects.filter(name=request.data['name']).exists():
                 update_data.save()
                 return Response(update_data.data, status=status.HTTP_200_OK)
             else:
@@ -72,17 +71,23 @@ class ReadUpdateDeleteView(APIView):
         # Return errors if the data is invalid
         return Response(update_data.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, user_id, *args, **kwargs):
+    def delete(self, request, user_id=None, name=None, *args, **kwargs):
+        # Check to confirm that either user_id or name is supplied to the view
+        if user_id == None and name == None:
+            return Response({
+                "message": "You must supply a user_id or name with this method.",
+            }, status=status.HTTP_400_BAD_REQUEST)
+
         # Read info aobut a person using id or name
-        name = request.query_params.get('name')
         if name:
             person = get_object_or_404(Person, name=name)
-        else:
+            person.delete()
+            return Response({'message': 'Person deleted.'})
+            
+        elif user_id:
             person = get_object_or_404(Person, pk=user_id)
-        person.delete()
-        return Response({
-            "message": f"Person with id {user_id} has been deleted"
-        }, status=status.HTTP_204_NO_CONTENT)
-
-base_view = BasePersonView.as_view()
-rud_view = ReadUpdateDeleteView.as_view()
+            person.delete()        
+            return Response({'message': 'Person deleted.'})
+        
+base_view = BaseView.as_view()
+rud_view = CRUDView.as_view()
